[MOVE] : musical_detail_image 패키지 안의 파일들 musical 패키지로 합침
- MusicalDetailImage 는 다른 엔티티들과 연관되어있지 않고 오직 Musical 에만 종속되기 때문
