+++ src/main/java/com/prgrms/be/intermark/domain/musical/repository/MusicalDetailImageRepository.java
@@ -1,10 +1,10 @@
-package com.prgrms.be.intermark.domain.musical_image.repository;
+package com.prgrms.be.intermark.domain.musical.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.prgrms.be.intermark.domain.musical_image.model.MusicalDetailImage;
+import com.prgrms.be.intermark.domain.musical.model.MusicalDetailImage;
 
 public interface MusicalDetailImageRepository extends JpaRepository<MusicalDetailImage, Long> {
